requirements.txt
anthropic>=0.54.0
python-dotenv>=1.0.1
